UIS: fn_8016C614/fn_8016C674 take the event as u32 like fn_8016C5C4 (callers pass it unmasked); fn_8016A2D4 87.5->90.3, fn_8016A030 93.4->94.4, fn_80168918 95.8->96.5

+++ include/frontend/uistudio.h
@@ -335,9 +335,10 @@ f32* fn_8016C1A4(s32 n20, UISNodeInfo* pInfo);
 s8 fn_8016C270(UIStudio* pStudio, UISScreen* pScreen, UISNodeInfo* pInfo, UISWordStack* pStack, u8* pScript,
                u32 nArgs, const s32* pArgs, u32 nArgs2, const s32* pArgs2, u8 bExtra, s32 nExtra, s32* pnSaved);
 // A node's handler scripts for an event, by kind (0x4000, plain with an ID, 0x8000); NULL for none.
-u8* fn_8016C5C4(UISNode* pNode, u32 uEvent);   // callers pass the event unmasked
-u8* fn_8016C614(UISNode* pNode, u16 uId, u16 uEvent);
-u8* fn_8016C674(UISNode* pNode, u16 uEvent);
+// The event is a u32 (callers pass it unmasked; each function masks it to 16 bits).
+u8* fn_8016C5C4(UISNode* pNode, u32 uEvent);
+u8* fn_8016C614(UISNode* pNode, u16 uId, u32 uEvent);
+u8* fn_8016C674(UISNode* pNode, u32 uEvent);
 u16 fn_8016C6C4(UIStudio* pStudio, u16 uGroup, u16 uScreen);
 
 // Sends event uEvent to the current screen, or to every screen when bAll is set; n -8 skips a
