+++ src/UISScreen.c
@@ -800,11 +800,11 @@ u8* fn_8016C5C4(UISNode* pNode, u32 uEvent) {
 }
 
 // A node's plain handler (neither kind bit) with the given ID for an event.
-u8* fn_8016C614(UISNode* pNode, u16 uId, u16 uEvent) {
+u8* fn_8016C614(UISNode* pNode, u16 uId, u32 uEvent) {
     u32 i;
     for (i = 0; i < pNode->nHandlers; i++) {
         UISHandler* pHandler = &pNode->pHandlers[i];
-        if (!(pHandler->uFlags & 0xC000) && pHandler->uEvent == uEvent &&
+        if (!(pHandler->uFlags & 0xC000) && pHandler->uEvent == (u16)uEvent &&
             (pHandler->uFlags & 0x2FFF) == uId) {
             return pHandler->u4.pScript;
         }
@@ -813,11 +813,11 @@ u8* fn_8016C614(UISNode* pNode, u16 uId, u16 uEvent) {
 }
 
 // A node's handler of the kind marked 0x8000 for an event.
-u8* fn_8016C674(UISNode* pNode, u16 uEvent) {
+u8* fn_8016C674(UISNode* pNode, u32 uEvent) {
     u32 i;
     for (i = 0; i < pNode->nHandlers; i++) {
         UISHandler* pHandler = &pNode->pHandlers[i];
-        if ((pHandler->uFlags & 0x8000) && pHandler->uEvent == uEvent) {
+        if ((pHandler->uFlags & 0x8000) && pHandler->uEvent == (u16)uEvent) {
             return pHandler->u4.pScript;
         }
     }
